Add script to fetch from Strava
I've written a bit of Python that uses the Strava API to create a
JSON file of commute data that matches the format that the
Processing code expects.

This also adds a temporary "spacer" element to Processing, so that
the output looks good until I've separated things out into morning
and evening commutes.